Fixes #56 - refactoring Login component
