+++ frontend/src/App.js
@@ -1,6 +1,6 @@
 import './App.css';
 import Home from './components/Home';
-import Login from './components/Login';
+import Login from './components/Login/Login';
 import Register from './components/Register';
 import React from "react";
 import {BrowserRouter as Router, Route,Routes} from 'react-router-dom';
